Refactor: move flatten_parallel_tool_calls to utils and merge tests
- Move flatten_parallel_tool_calls from generic.py to OCIUtils in utils.py
- Merge unit tests into test_oci_generative_ai.py
- Merge integration tests into test_tool_calling.py
- Delete standalone test files

Addresses reviewer feedback on PR #111.
